+++ Game/src/menus/meter.cpp
@@ -60,7 +60,7 @@ void Meter::Init( int initialValue, int maxValue )
 
 void Meter::SetValue( int value )
 {
-	for ( int i = 0, c = m_Panels.size(); i < c; ++i )
+	for ( size_t i = 0, c = m_Panels.size(); i < c; ++i )
 	{
 		m_Panels[ i ]->Show( ( i + 1 ) <= value );
 	}
