+++ Game/src/ship/weapon.cpp
@@ -401,7 +401,7 @@ void Weapon::AddMuzzle( const glm::vec3& position )
 
 void Weapon::MarkMuzzleAsUsed()
 {
-	unsigned int validMuzzles = m_MuzzleOffsets.size();
+	unsigned int validMuzzles = static_cast<unsigned int>(m_MuzzleOffsets.size());
 	if ( validMuzzles > 1 && ++m_CurrentMuzzle >= validMuzzles )
 	{
 		m_CurrentMuzzle = 0;
